requirements.txt
Flask>=3.0.0,<4.0.0
numpy>=1.20.0,<2.0.0
anakin-language-server>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
scikit-learn>=1.2.0,<1.3.0
black>=24.0.0,<25.0.0
pandas>=2.2.0,<2.3.0
python-dotenv>=1.0.0,<2.0.0
